requirements.txt

streamlit
geopandas
pandas
leafmap
Pillow
pyogrio
fiona
